git-svn-id: https://svn.apache.org/repos/asf/apr/apr-util/branches/1.7.x@1925641 13f79535-47bb-0310-9956-ffa450edef68

+++ test/testpass.c
@@ -125,7 +125,7 @@ static void test_shapass(abts_case *tc, void *data)
     const char *pass2 = "hellojed2";
     char hash[100];
 
-    apr_sha1_base64(pass, strlen(pass), hash);
+    apr_sha1_base64(pass, (unsigned int)strlen(pass), hash);
 
     apr_assert_success(tc, "SHA1 password validated",
                        apr_password_validate(pass, hash));
